reviewfix: Made LazyStringValue static inner class of  CollapsingQParserPlugin
